Made tests
I'm new to tests in py, I don't know how to do them. I'm sorry to everyone's eyes.
