+++ tests/important.md
@@ -1,0 +1,4 @@
+You must run starter.py before running initiating tests as stats are only generated then.
+You don't have to run starter.py again if you have already run it.
+
+you do not have to play through the chapters. Just wait until you see "Welcome to melancholy!" Then press CTRL + C to exit and run the tests.
